move all functions related to running benchmarks to csbench_run.c
